fix call-during-call: switch off re-using lchan during assignment
Disable lchan re-use on repeated Assignment Requests from the MSC, in order to
enable switching over to a new call that is coming in during an ongoing call.

This disables a rare and untested scenario for the benefit of a very common
one: call during call. It adds a feature by disabling code :)

The long-term way to support call-during-call, as well as Channel Mode Modify,
would be to rather re-use the old lchan and implement FSM states to re-route
the audio. But that is a task for another day.

A more detailed explanation is given in the '#if 0' section added by this
patch.

To really get call-during-call working, osmo-msc also needs a fix, see the
change-id in 'Related' below.

Related: OS#3735, osmo-msc I0ba216b737909e92080a722db26e3577726c63cb
Change-Id: I0c00ec2c120e5008281755adcd4944a3ce4d8355