manual: It's not "A over SCCP" but "BSSAP over SCCP"
In GSM specs, the entire interface between two elements is designated
with some letter, like the A interface between BSC and MSC.  The
interface uses a variety of protocols stacked on each other.

In the specific case of A, there is no "A" on top of SCCP, but
there's "BSSAP" on top of SCCP.

This is followed somewhat un-orthodox by 3GPP, as "A over IP" is
a violation of that principle.  It should have been called "A utilizing
IP", "A based on IP", "A with IP" or something the like.

In any case, at no point do the specs ever claim that "A" is stacked
on top of SCCP, so let's fix this.

Change-Id: Ieb0d8f6c71debe1234aff343a994c2096326da1b