osmo-bsc/Makefile.am: use LIBOSMOLEGACYMGCP_CFLAGS
Fix the following compile error, when libosmo-legacy-mgcp is installed,
but into a different prefix than the other libosmo* libraries:

In file included from osmo_bsc_msc.c:23:0:
../../include/osmocom/bsc/bsc_nat.h:24:38: fatal error: osmocom/legacy_mgcp/mgcp.h: No such file or directory
 #include <osmocom/legacy_mgcp/mgcp.h>

This is needed for a new CI job, which builds old Osmocom releases
against current master of all Osmocom libraries. libosmo-legacy-mgcp is
provided by osmo-mgw 1.4.0 and was removed in later versions, so we
want to have the 1.4.0 version installed into a fallback directory only.

Related: OS#3867
Change-Id: Ibd7948f12da710f8ca2b8fde8870f134308eb908