chan_alloc: Don't include CBCH channels in load computation
When displaying the channel (lchan) load, we are interested in the
dedicated channels as available for communication with mobile
stations.  If CBCH is enabled, it "steals" one of the SDCCH slots,
so we have to exclude this channel from the count of available
dedicated channels.

Before this patch, A CBCH would show up as "in use" dedicated channel
(so both the total and the in_use count were incremented).

After this patch, neither "total" nor "use" are affected; The CBCH
slot is completely excluded from load statistics.

Change-Id: I606962e4977f122ff05a2d4e756c04c7f38bdaa0