lchan_fsm: allow transition from broken state to release
If a channel activation ack comes in late from a BTS, the borken state has a
transition back to a channel recovery. This only fails because of a missing
state transition in the lchan_fsm. Allow the transition.

We've seen this happening at cccamp2019, where a particular BTS for unknown
reasons sends its Chan Act ACK a second after the activation times out in
osmo-bsc. This patch should allow it to recover.

Change-Id: I73139033ae44abb383ccce956c1c9683db8c8816