a_reset.c: Don't wait 2 seconds to send first BSSMAP RESET
Function fsm_reset_ack_timeout_cb() could be called directly from within
a_reset_alloc(), but it's still desirable to deferr the BSSMAP RESET to
be sent asynchronously by the timer upon next main loop step as soon as
possible, so whole process is already configured properly.

1ms needs to be set instead of 0 (immediate asynchronous) because value
0 actually disables the timer.

As a result, moving the state_chg() after the msc->a.reset_fsm
assignment is not really needed, but still makes it more clear that the
pointer will be set upon call of the timer callback.

Related: OS#4188
Change-Id: I68d76a4050d4dec7d53b0031d67e0dd35ddd8764