BS-11: Fix "CONNECT TERRESTRIAL TRAFFIC"
On a TS 12.21 spec compliant BTS, each Um traffic channel must be mapped
to an E1 sub-slot on the terrestrial back-haul.  This happens via
the CONNECT TERRESTRIAL TRAFFIC message.

We always had code to sen that message, but it got deactivated when
ts->pchan_is / ts->pchan_from_config was introduced.  Of course,
while we are bringing up OML, there is no 'pchan_is' set yet. We only
have 'pchan_from_config' and must use it.

Change-Id: I8988a027b0e897bd9dda460590f974d6be34a4fa