abis_rsl: prioritize emergency calls over regular calls
when an emergency call arrives while all TCH are busy, the BSC should
pick an arbitrary (preferably the longest lasting) call / lchan and
release it in favor of the incoming emergancy call.

The release of the existing call is a process that can not be done
synchronously while the ChanRQD is handled sonce multiple messages are
exchanged between BTS and MSC and multiple FSMs need to do their work.

To be able to release one lchan while handling a ChanRQD a queue is
implemented in which the incomming channel requests are collected. The
queue is checked regulary for incoming requests. If one of the requests
is for an emergency call and if all channels are busy one channel is
picked and released. It is also made sure that the incoming emergency
call directly gets the TCH that was just released.

Change-Id: If8651265928797dbda9f528b544931dcfa4a0b36
Related: OS#4549